+++ src/ATK/configuration/defaults/base_defaults.py
@@ -14,7 +14,7 @@ DEFAULTS = {
         "font_size": 14,
         "font": "Helvetica",
     },
-    "query_settings": {"query_radius": 3, "image_size": 30, "default_unit": "arcsec"},
+    "query_settings": {"query_radius": 3, "image_size": 30, "default_scale": "arcsec"},
     "overlay_settings": {"piggyback_radius": 5, "simbad_radius": 3},
     "search_settings": {"search_radius": 3},
     "datapage_settings": {"search_button_radius": 3, "datatable_radius": 3, "font_size": 12, "grid_size": 250},